refactor: consolidate engine enums into model/enums.ts
Move every enum created in the string-literal refactor into a single home,
libs/engine-core/src/lib/model/enums.ts: Zone, CardType, SuperType, Keyword,
TargetKind, Phase, Step, GameStatus, StackItemSource, EffectType,
GameEventType, LifeChangeReason, PriorityResetReason, CombatDeclaration,
PlayerLostReason, and DamagePass.

The former owners re-export their relevant enums (model/types.ts re-exports
all of them; engine/events.ts re-exports the event enums; effect-types.ts
re-exports EffectType) so existing import paths keep resolving — no consumer
churn beyond the two StackItemSource importers, repointed to model/types.

Left in place: ActionType (pre-existing, bound to the Action union) and the
protocol message enums. ConnectionStatus stays in game-ui — it is a client
transport concept, not a game-domain model type, and engine-core is pure.

Build, lint, format, and the full vitest suite pass.

[CLAUDE]
