+++ libs/engine-core/src/lib/engine/validators/cast-creature.validator.ts
@@ -4,8 +4,7 @@ import type { CastCreature } from '../../actions/action';
 import { getCardDefinition } from '../../cards/catalog';
 import type { GameState } from '../../model/game-state';
 import type { StackItem } from '../../model/stack';
-import { StackItemSource } from '../../model/stack';
-import { CardType, Zone } from '../../model/types';
+import { CardType, StackItemSource, Zone } from '../../model/types';
 import type { GameEvent } from '../events';
 import { GameEventType } from '../events';
 import { manaSpentMatchesCost, poolHasAtLeast, totalSpent } from '../mana';
