+++ libs/engine-core/src/lib/engine/validators/cast-non-permanent-spell.ts
@@ -5,8 +5,7 @@ import { getCardDefinition } from '../../cards/catalog';
 import { EffectType } from '../../cards/effects/effect-types';
 import type { GameState } from '../../model/game-state';
 import type { StackItem } from '../../model/stack';
-import { StackItemSource } from '../../model/stack';
-import { type CardType, Zone } from '../../model/types';
+import { StackItemSource, Zone, type CardType } from '../../model/types';
 import type { GameEvent } from '../events';
 import { GameEventType } from '../events';
 import { manaSpentMatchesCost, poolHasAtLeast } from '../mana';
